Check if the image location is null

+++ app/src/main/java/com/hci_capstone/poison_ivy_tracker/poison_ivy_reporter/database/ReportDatabase.java
@@ -143,8 +143,10 @@ public abstract class ReportDatabase extends RoomDatabase {
                     continue;
                 }
                 for (String imageLocation : report.getImageLocations()) {
-                    File image = new File(imageLocation);
-                    image.delete();
+                    if (imageLocation != null) {
+                        File image = new File(imageLocation);
+                        image.delete();
+                    }
                 }
             }
